:recycle: Refactor: Converter 리팩토링 (#30)
