--- src/main/java/site/moamoa/backend/converter/post/PostConverter.java
@@ -1,51 +1,0 @@
-package site.moamoa.backend.converter.post;
-
-import java.util.List;
-import site.moamoa.backend.domain.Category;
-import site.moamoa.backend.domain.Post;
-import site.moamoa.backend.domain.enums.CapacityStatus;
-import site.moamoa.backend.domain.mapping.PostImage;
-import site.moamoa.backend.web.dto.request.PostRequestDTO.AddPost;
-import site.moamoa.backend.web.dto.response.PostResponseDTO.AddPostResult;
-import site.moamoa.backend.web.dto.response.PostResponseDTO.UpdatePostInfoResult;
-import site.moamoa.backend.web.dto.response.PostResponseDTO.UpdatePostStatusResult;
-
-public class PostConverter {
-
-    public static Post toPost(AddPost addPost, Category category, List<PostImage> postImages){
-        return Post.builder()
-            .available(addPost.personnel())
-            .viewCount(0)
-            .capacityStatus(CapacityStatus.NOT_FULL)
-            .category(category)
-            .personnel(addPost.personnel())
-            .deadline(addPost.deadline())
-            .productName(addPost.productName())
-            .postImages(postImages)
-            .dealLocation(addPost.dealLocation())
-            .totalPrice(addPost.price())
-            .description(addPost.description())
-            .build();
-    }
-
-    public static AddPostResult toAddPostResult(Post post){
-        return AddPostResult.builder()
-            .postId(post.getId())
-            .createdAt(post.getCreatedAt())
-            .build();
-    }
-
-    public static UpdatePostInfoResult toUpdatePostInfoResult(Post post) {
-        return UpdatePostInfoResult.builder()
-            .postId(post.getId())
-            .updatedAt(post.getUpdatedAt())
-            .build();
-    }
-
-    public static UpdatePostStatusResult toUpdatePostStatusResult(Post post) {
-        return UpdatePostStatusResult.builder()
-            .postId(post.getId())
-            .updatedAt(post.getUpdatedAt())
-            .build();
-    }
-}
